# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for book order landing page
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Wilaya {
  id: number;
  name: string;
  zone: number;
  is_deliverable: number;
}

export interface Commune {
  id: number;
  name: string;
  wilaya_id: number;
  wilaya_name?: string;
  has_stop_desk: number;
  is_deliverable: number;
}

export interface Center {
  center_id: number;
  name: string;
  commune_id?: number;
  commune_name?: string;
  wilaya_id?: number;
  wilaya_name?: string;
}

export interface OrderInput {
  /** @minLength 1 */
  firstname: string;
  /** @minLength 1 */
  familyname: string;
  /** @minLength 9 */
  contact_phone: string;
  /** @nullable */
  address?: string | null;
  to_wilaya_name: string;
  to_commune_name: string;
  is_stopdesk: boolean;
  /** @nullable */
  stopdesk_id?: number | null;
}

export interface OrderResult {
  success: boolean;
  /** @nullable */
  tracking: string | null;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  message?: string | null;
}

export type GetCommunesParams = {
wilaya_id: number;
};

export type GetCentersParams = {
wilaya_id: number;
};