# lib/api-zod/src/generated/types/center.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for book order landing page
 * OpenAPI spec version: 0.1.0
 */

export interface Center {
  center_id: number;
  name: string;
  commune_id?: number;
  commune_name?: string;
  wilaya_id?: number;
  wilaya_name?: string;
}