# lib/api-zod/src/generated/types/commune.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for book order landing page
 * OpenAPI spec version: 0.1.0
 */

export interface Commune {
  id: number;
  name: string;
  wilaya_id: number;
  wilaya_name?: string;
  has_stop_desk: number;
  is_deliverable: number;
}