# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for book order landing page
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  /** @minLength 1 */
  firstname: string;
  /** @minLength 1 */
  familyname: string;
  /** @minLength 9 */
  contact_phone: string;
  /** @nullable */
  address?: string | null;
  to_wilaya_name: string;
  to_commune_name: string;
  is_stopdesk: boolean;
  /** @nullable */
  stopdesk_id?: number | null;
}