# lib/api-zod/src/generated/types/orderResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for book order landing page
 * OpenAPI spec version: 0.1.0
 */

export interface OrderResult {
  success: boolean;
  /** @nullable */
  tracking: string | null;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  message?: string | null;
}