# lib/api-zod/src/generated/types/wilaya.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for book order landing page
 * OpenAPI spec version: 0.1.0
 */

export interface Wilaya {
  id: number;
  name: string;
  zone: number;
  is_deliverable: number;
}